Add JSON logging support (#3305)
* Add JSON logging support

This adds support for JSON logging, along with the relevant options required.

This does not change the default log behaviour, so should be backwards compatible.

It is opt in via the LOG_FORMAT option, which can be 'json' to use the new logger, or anything else to fallback
to the old behaviour.

This is implemented in terms of a custom formatter, which is optionally applied to the stdout stream.
The debug stream is unaffected by this.

* Allow JSON to be selected when creating handlers

* Allow different formats to be selected for streams/files

* Remove old / unused code

* Add new config opts to helpfile

* Formatting, basic typing and reorder for consistency in project.

---------

Co-authored-by: Jerrie-Aries <hidzrie@gmail.com>
Co-authored-by: Taku <45324516+Taaku18@users.noreply.github.com>

(cherry picked from commit 6d61cf29ff0d9e8a3b984eb44a6337402246bff0)
Signed-off-by: Khakers <22665282+khakers@users.noreply.github.com>